+++ src/tests/unit/common/utils/test_build_mongo_query.py
@@ -1,7 +1,9 @@
 import unittest
 
-from common.utils.build_complex_search import build_mongo_query
 from domain_classes.blueprint import Blueprint
+from features.search.use_cases.search_use_case.build_complex_search import (
+    build_mongo_query,
+)
 
 basic_blueprint = {
     "type": "system/SIMOS/Blueprint",
